save results from main; stratified perf metrics; causal model validation: CCA; fix args type in parser

+++ src/utils.py
@@ -16,7 +16,7 @@ def parse_args():
 
   parser.add_argument('--batch_size', type=int, default=Config.BATCH_SIZE)
   parser.add_argument('--n_epochs', type=int, default=Config.N_EPOCHS)
-  parser.add_argument('--lr', type=int, default=Config.LEARNING_RATE)
+  parser.add_argument('--lr', type=float, default=Config.LEARNING_RATE)
   parser.add_argument('--distill_kl_ann', type=int, default=Config.DISTILL_KL_ANN)
 
   parser.add_argument('--seed', type=int, default=Config.SEED)
@@ -24,11 +24,11 @@ def parse_args():
   parser.add_argument('--ud_dim', type=int, default=Config.UD_DIM)
   parser.add_argument('--h_dim', type=int, default=Config.H_DIM)
   parser.add_argument('--act_fn', type=str, default=Config.ACT_FN)
-  parser.add_argument('--corr_a', type=int, default=Config.CORR_RECON_ALPHA)
-  parser.add_argument('--desc_a', type=int, default=Config.DESC_RECON_ALPHA)
-  parser.add_argument('--pred_a', type=int, default=Config.PRED_ALPHA)
-  parser.add_argument('--fair_b', type=int, default=Config.FAIR_BETA)
-  parser.add_argument('--tc_b', type=int, default=Config.TC_BETA)
+  parser.add_argument('--corr_a', type=float, default=Config.CORR_RECON_ALPHA)
+  parser.add_argument('--desc_a', type=float, default=Config.DESC_RECON_ALPHA)
+  parser.add_argument('--pred_a', type=float, default=Config.PRED_ALPHA)
+  parser.add_argument('--fair_b', type=float, default=Config.FAIR_BETA)
+  parser.add_argument('--tc_b', type=float, default=Config.TC_BETA)
 
   parser.add_argument('--exp_name', type=str, default=date_str)
